2.x: Factor out typical TestObserver/TestSubscriber methods
